config-file: reorganize SSL options
This is preparation for a new authentication method mTLS being
introduced in a future commit.

Move the SSL options up, so mTLS options can be added before bailing out
due to no valid authentication option set. Also move the existence check
for auth/gateway token into a new variable, so we can check for them in
a combined fashion.

Signed-off-by: Robin van der Gracht <robin@protonic.nl>
Signed-off-by: Bastian Krause <bst@pengutronix.de>